better nesting for data/api objects
