+++ lib/trustly.rb
@@ -14,7 +14,7 @@ require 'trustly/exception/configuration_error'
 require 'trustly/exception/jsonrpc_version_error'
 require 'trustly/exception/signature_error'
 
-require 'trustly/data'
+require 'trustly/data/base'
 require 'trustly/data/request'
 require 'trustly/data/response'
 require 'trustly/data/jsonrpc_request'
@@ -22,6 +22,6 @@ require 'trustly/data/jsonrpc_response'
 require 'trustly/data/jsonrpcnotification_request'
 require 'trustly/data/jsonrpcnotification_response'
 
-require 'trustly/api'
+require 'trustly/api/base'
 require 'trustly/api/signed'
 require 'trustly/version'
