+++ lib/trustly/data/jsonrpc_response.rb
@@ -1,17 +1,21 @@
-class Trustly::Data::JSONRPCResponse < Trustly::Data::Response
-  VERSION_ERROR = 'JSON RPC Version is not supported'
+ module Trustly
+  module Data
+    class JSONRPCResponse < Response
+      VERSION_ERROR = 'JSON RPC Version is not supported'
 
-  def initialize(**options)
-    super
-    version = payload['version']
-    if version != '1.1'
-      raise Trustly::Exception::JSONRPCVersionError, VERSION_ERROR 
+      def initialize(**options)
+        super
+        version = payload['version']
+        if version != '1.1'
+          raise Trustly::Exception::JSONRPCVersionError, VERSION_ERROR 
+        end
+      end
+
+      def data_at(name)
+        return if data.nil?
+        
+        data[name]
+      end
     end
-  end
-
-  def data_at(name)
-    return if data.nil?
-    
-    data[name]
   end
 end
